Conflicts:
	src/Tweetsharp.PCL/Serialization/JsonSerializer.cs

+++ src/Tweetsharp.PCL/TwitterService.OAuth.cs
@@ -231,7 +231,7 @@ namespace TweetSharp
             {
                 Token = query["oauth_token"] ?? "?",
                 TokenSecret = query["oauth_token_secret"] ?? "?",
-                UserId = Convert.ToInt32(query["user_id"] ?? "0"),
+                UserId = Convert.ToInt64(query["user_id"] ?? "0"),
                 ScreenName = query["screen_name"] ?? "?"
             };
 
@@ -264,7 +264,7 @@ namespace TweetSharp
                         {
                             Token = query["oauth_token"] ?? "?",
                             TokenSecret = query["oauth_token_secret"] ?? "?",
-                            UserId = Convert.ToInt32(query["user_id"] ?? "0"),
+                            UserId = Convert.ToInt64(query["user_id"] ?? "0"),
                             ScreenName = query["screen_name"] ?? "?"
                         };
 
@@ -352,7 +352,7 @@ namespace TweetSharp
                                             {
                                                 Token = query["oauth_token"] ?? "?",
                                                 TokenSecret = query["oauth_token_secret"] ?? "?",
-                                                UserId = Convert.ToInt32(query["user_id"] ?? "0"),
+                                                UserId = Convert.ToInt64(query["user_id"] ?? "0"),
                                                 ScreenName = query["screen_name"] ?? "?"
                                             };
                                             return accessToken;
@@ -397,7 +397,7 @@ namespace TweetSharp
                                        {
                                            Token = query["oauth_token"] ?? "?",
                                            TokenSecret = query["oauth_token_secret"] ?? "?",
-                                           UserId = Convert.ToInt32(query["user_id"] ?? "0"),
+                                           UserId = Convert.ToInt64(query["user_id"] ?? "0"),
                                            ScreenName = query["screen_name"] ?? "?"
                                        };
                                        return accessToken;
